move TS defs into separate src/lib/types.ts file
